requirements.txt
nltk==3.1
numpy==1.10.2
oauthlib==1.0.3
python-twitter==2.2
requests==2.9.1
requests-oauthlib==0.6.0
scikit-learn==0.17
scipy==0.16.1
